+++ workflow/controller/indexes/indexes.go
@@ -3,6 +3,7 @@ package indexes
 const (
 	ClusterWorkflowTemplateIndex = "clusterworkflowtemplate"
 	CronWorkflowIndex            = "cronworkflow"
+	WorkflowIndex                = "workflow"
 	WorkflowTemplateIndex        = "workflowtemplate"
 	WorkflowPhaseIndex           = "workflow.phase"
 	SemaphoreConfigIndexName     = "bySemaphoreConfigMap"
