+++ workflow/controller/indexes/labels.go
@@ -3,12 +3,8 @@ package indexes
 import (
 	"fmt"
 
-	"github.com/prometheus/common/log"
 	"k8s.io/apimachinery/pkg/api/meta"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/client-go/tools/cache"
-
-	"github.com/argoproj/argo/workflow/util"
 )
 
 func MetaNamespaceLabelIndex(namespace, label string) string {
@@ -26,22 +22,6 @@ func MetaLabelIndexFunc(label string) cache.IndexFunc {
 		} else {
 			return []string{}, nil
 		}
-	}
-}
-
-func WorkflowSemaphoreKeysIndexFunc() cache.IndexFunc {
-	return func(obj interface{}) ([]string, error) {
-		un, ok := obj.(*unstructured.Unstructured)
-		if !ok {
-			log.Warnf("cannot convert obj into unstructured.Unstructured in Indexer %s", SemaphoreConfigIndexName)
-			return []string{}, nil
-		}
-		wf, err := util.FromUnstructured(un)
-		if err != nil {
-			log.Warnf("failed to convert to workflow from unstructured: %v", err)
-			return []string{}, nil
-		}
-		return wf.GetSemaphoreKeys(), nil
 	}
 }
